requirements.txt
PyQt6>=6.4.0
pywin32>=305
psutil>=5.9.0
Pillow>=9.0.0